// src/app/interfaces/porcinoSumary.ts
export interface PorcinoSumary {
    idPorcino: string;
    edad: number;
    peso: number;
    idCliente: string;
    nombres: string;
    apellidos: string;
    direccion: string;
    telefono: string;
    idRaza: string;
    descripcionRaza: string;
    idAlimentacion: string;
    descripcionAl: string;
    dosis: string;
}